+++ lib/core/nest-application.d.ts
@@ -7,6 +7,7 @@ import { ApplicationConfig } from './application-config';
 import { NestContainer } from './injector/container';
 import { NestApplicationContext } from './nest-application-context';
 import { NestApplicationOptions } from '@nestjs/common/interfaces/nest-application-options.interface';
+import { CorsOptions } from '@nestjs/common/interfaces/external/cors-options.interface';
 export declare class NestApplication extends NestApplicationContext implements INestApplication {
     private readonly httpAdapter;
     private readonly config;
@@ -20,9 +21,14 @@ export declare class NestApplication extends NestApplicationContext implements I
     private readonly microservices;
     private httpServer;
     private isInitialized;
+<<<<<<< HEAD
     constructor(container: NestContainer, httpAdapter: any, config: ApplicationConfig, appOptions?: NestApplicationOptions);
     registerHttpServer(): void;
     applyOptions(): any;
+=======
+    constructor(container: NestContainer, express: any, config: ApplicationConfig, appOptions?: NestApplicationOptions);
+    applyOptions(): this;
+>>>>>>> master
     createServer(): any;
     getUnderlyingHttpServer(): any;
     registerModules(): Promise<void>;
@@ -40,7 +46,7 @@ export declare class NestApplication extends NestApplicationContext implements I
     set(...args: any[]): this;
     disable(...args: any[]): this;
     enable(...args: any[]): this;
-    enableCors(): this;
+    enableCors(options?: CorsOptions): this;
     listen(port: number | string, callback?: () => void): any;
     listen(port: number | string, hostname: string, callback?: () => void): any;
     listenAsync(port: number | string, hostname?: string): Promise<any>;