+++ src/common/exceptions/http.exception.ts
@@ -1,9 +1,14 @@
-export class HttpException {
-  private readonly message: string | object;
+export class HttpException extends Error {
+  public readonly message: any;
 
   /**
+<<<<<<< HEAD
    * The base Nest Application exception which is handled by the default Exceptions Handler.
    * If you throw an exception from your HTTP route handlers, Nest will map them to the appropriate HTTP response and send to the client.
+=======
+   * The base Nest Application exception, which is handled by the default Exceptions Handler.
+   * If you throw an exception from your HTTP route handler, Nest will map them to the appropriate HTTP response and send to the client.
+>>>>>>> master
    *
    * When `response` is an object:
    * - object will be stringified and returned to the user as a JSON response,
@@ -19,6 +24,7 @@ export class HttpException {
     private readonly response: string | object,
     private readonly status: number,
   ) {
+    super();
     this.message = response;
   }