+++ src/common/interfaces/nest-application.interface.ts
@@ -3,6 +3,7 @@ import { WebSocketAdapter } from './websockets/web-socket-adapter.interface';
 import { CanActivate } from './features/can-activate.interface';
 import { NestInterceptor } from './features/nest-interceptor.interface';
 import { INestApplicationContext } from './nest-application-context.interface';
+import { CorsOptions } from './external/cors-options.interface';
 
 export interface INestApplication extends INestApplicationContext {
   /**
@@ -49,7 +50,7 @@ export interface INestApplication extends INestApplicationContext {
    *
    * @returns void
    */
-  enableCors(): this;
+  enableCors(options?: CorsOptions): this;
 
   /**
    * A wrapper function around native `express.disable()` method.