+++ src/core/injector/module.ts
@@ -50,8 +50,8 @@ export class Module {
   private _exports = new Set<string>();
 
   constructor(
-    private _metatype: Type<any>,
-    private _scope: Type<any>[],
+    private readonly _metatype: Type<any>,
+    private readonly _scope: Type<any>[],
     container: NestContainer,
   ) {
     this.addCoreInjectables(container);
@@ -145,9 +145,9 @@ export class Module {
       metatype: ExternalContextCreator,
       isResolved: true,
       instance: new ExternalContextCreator(
-        new GuardsContextCreator(container),
+        new GuardsContextCreator(container, container.applicationConfig),
         new GuardsConsumer(),
-        new InterceptorsContextCreator(container),
+        new InterceptorsContextCreator(container, container.applicationConfig),
         new InterceptorsConsumer(),
         container.getModules(),
       ),
@@ -271,11 +271,9 @@ export class Module {
   ) {
     if (this.isCustomProvider(exportedComponent as any)) {
       return this.addCustomExportedComponent(exportedComponent as any);
-    }
-    else if (isString(exportedComponent)) {
+    } else if (isString(exportedComponent)) {
       return this._exports.add(exportedComponent);
-    }
-    else if (this.isDynamicModule(exportedComponent)) {
+    } else if (this.isDynamicModule(exportedComponent)) {
       const { module } = exportedComponent;
       return this._exports.add(module.name);
     }